Move Translation Panel to extra/Locale
