+++ meerk40t/gui/wxmeerk40t.py
@@ -5,7 +5,6 @@ import sys
 import traceback
 
 from meerk40t.gui.wxmscene import SceneWindow
-from .translation import Translation
 
 try:
     import wx
@@ -291,7 +290,6 @@ class wxMeerK40t(wx.App, Module):
         kernel.register("window/RasterWizard", RasterWizard)
         kernel.register("window/Simulation", Simulation)
         kernel.register("window/Scene", SceneWindow)
-        kernel.register("window/Translate", Translation)
 
         kernel.register("window/default/Controller", Controller)
         kernel.register("window/default/Preferences", Configuration)
